+++ data-raw/internal_data.R
@@ -1,6 +1,6 @@
 #===========================statistic_index================
 #=============possible calculated statistics===============
-single.stats <- list(stat = c("ho", "pi", "he", "maf", "private", "association", "hwe", "random_forest", "genomic_prediction", "fis"),
+single.stats <- list(stat = c("ho", "pi", "he", "maf", "private", "association", "hwe", "random_forest", "genomic_prediction", "fis", "allelic_richness"),
                      col_pattern = list("ho", 
                                 "pi",
                                 "he",
@@ -10,7 +10,8 @@ single.stats <- list(stat = c("ho", "pi", "he", "maf", "private", "association",
                                 c("pHWE"),
                                 c("RF_importance", "RF_importance_pvals"),
                                 c("gp_effect"),
-                                c("fis", "var_comp_b", "var_comp_c", "nk")))
+                                c("fis", "var_comp_b", "var_comp_c", "nk"),
+                                c("richness")))
 window.stats <- list(stat = "tajimas_d",
                   col_pattern = list(c("ws.theta", "ts.theta", "num_seg", "D", "n_snps")))
 pairwise.stats <- list(stat = c("fst", "abba_baba"),